Change configuration for maximum starman worker process memory size
Previously this was 300MB. But the virtual and resident sizes for a single
worker process are both significantly larger. Because we were checking it
every 10 requests, the worker processes were restarted after serving 10
requests.

Change it to 1300MB so that it is slightly larger than the virtual size for
a single worker which is 1260MB minimum.