Add Map controller to eg-rest as a temporary solution
While this pull request (https://github.com/Ensembl/ensembl-rest/pull/161) is
being worked on, add the changes to the eg-rest so that it could override the
controller on ensembl-rest.

This is a temporary solution. After the above pull request is accepted this
change should no longer be required and hence the Map controller should be
removed from eg-rest.