Add config for avoiding long db lookup
It is now possible to avoid long db lookup for ids that do not exist in the stabe_ids
database with this change: https://github.com/Ensembl/ensembl-rest/commit/59088c99e71201cecee7b8b6cb4ac82b165de822

Previously if an id does not exist in the stable_ids database, it takes very long time
for EG (>42k species) to iterate over each species id and look for that id. Recently
we have improve the populate_stable_ids script to include all the ids from multiple db
servers, we do not need to lookup an id from species db if that id is not found in the
stable_ids database.

Related to ENSEMBL-4859.