Add sending SIGQUIT first when stopping a Starman server
This is to ensure that the worker processes exit completely before the master
process is terminated. With recent change (5cc9b19), the master process shares
most of its memory with the worker processes. If the master process is killed
at the same time as the worker processes, all the shared memory will have to
be copied to those worker processes that are not yet terminated. This causes
long delay in stopping a Starman server.

Related to ENSEMBL-4873.