Look up genetree id from compara dbs
Previously we look the genetree id up first in the stable ids db. But the genetree
ids are not loaded onto stable ids db. We should look up the genetree id from compara
dbs directly.

This change removes the Lookup model in ensembl-rest and have a new copy of the Lookup
model in eg-rest with the fix in. This is a temporary solution. We should make the change
in ensembl-rest.

Fix ENSEMBL-4931.